moved all the field checks from the nodes isValid function to the fields isValid function
